// rc0/src/app/pages/MainPage/pages/PermissionPage/Sidebar/types.ts
/**
 * Seas
 *
 * Copyright 2021
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { ResourceTypes, SubjectTypes } from '../constants';

export interface PanelsProps {
  viewpointId: string | undefined;
  viewpointType: SubjectTypes | ResourceTypes | undefined;
  onToggle: (active: boolean, key: string) => void;
  onToDetail: (id: string, type: SubjectTypes | ResourceTypes) => () => void;
}